refactor: migrate to hexagonal architecture (FastAPI + Mangum, single HTTP Lambda)
Refactor UVA-App-Integrations to hexagonal architecture following the proven
pattern (DeviceDataAccess #23, ConnectionPrediction #8, UserAPI #12). All HTTP
endpoints are now served by ONE FastAPI + Mangum Lambda; the non-HTTP DynamoDB
Stream integrations stay as separate Lambdas sharing the same `src` build.

- src/core/use_cases: real logic moved verbatim (last_connection, create_racimo,
  dynamodb_to_sns, uva_to_cloud) — error branches preserved exactly.
- src/adapters/inbound/http: FastAPI app (routes, event_builder seam) with
  @app.exception_handler(Exception) -> 502 {"message": "Internal server error"}
  for API Gateway backend-error parity (e2e asserts 502).
- src/lambda_handlers: _bootstrap.py registers `src` alias for CodeUri: src;
  api_handler = Mangum(app); event_ingestor_* for the two streams.
- Legacy handlers kept as thin shims at exact path/name/signature (bootstrap
  sys.path + module-level import requests/boto3 for patch-by-reference +
  re-export of the use case) so integration tests pass unchanged.
- template.yaml: single HTTP function (logical ID UVALastConnection, both Api
  events) + two stream functions, all CodeUri: src, runtime python3.9.

Tests untouched (byte-identical). Integration: 40 passed. E2E local: 16 passed.
Deploy workflows (DeployMain.yml/DeployTest.yml) untouched.
