+++ SAM-UVA-App-Integrations/src/requirements.txt
@@ -1,0 +1,4 @@
+fastapi==0.110.0
+mangum==0.17.0
+requests
+botocore
